Conclave: enclave-host local communication (#233)
Conclave: enclave-host local communication

Introduced three new modules which will be part of the Conclave API: conclave-host, conclave-enclave and conclave-common.

conclave-host houses the new host API for bridging the enclave with the untrusted host. Only the local communication between the host and enclave is implemented in this commit. conclave-common contains that which is common between the two.

There are two main changes to the existing code:

1. NativeEnclaveApi has been extended to support Conclave by using Java's ServiceLoader API to avoid having a hard dependency to Conclave. This allows us to use the existing ECALL/OCALL JNI pathway.
2. Error handling in the handler framework has been improved to better support unchecked exceptions - they are now thrown as is rather than being wrapped again in another unchecked exception.