CON-33: EnclaveHost.loadFromResources
Implemented loadFromResources which looks up the enclave file on the classpath using the enclave class name, rather than having to specify the file. This means the host can be a single self-contained jar. To enable this the gradle plugin has been updated so that the enclave file is automatically in the right place on the host's classpath. The Conclave sample has been updated to showcase this.