CON-46: Removed Enclave-Class manifest attribute (#296)
With the EnclaveHost API requiring the enclave class name to load the enclave, the need for the Enclave-Class manifest attribute in the embedded enclave jar goes away. The load parameter instead can be fed in as the first ECALL to instantiate the Enclave object.

In the process, the gradle plugin tests can been cleaned up. There's only one test for checking an invalid encalve project. The other tests are removed as the hello world sample tests that better.