Introduced a single top-level plugin extension object, called "conclave" (#328)
The existing enclaveSimulation, enclaveDebug and enclaveRelease extension objects are now sub-configs and have the "enclave" prefix removed.

A common extension object is needed as there will be need of configuration which is mode-agnositic, and also seems to be gradle plugin best practice.

Also, tasks which need input from the extension objects now have the extension objects injected in, rather than having mirror properties. This gives better error messages to the user.