CON-68: Migrated tests using old enclave API to use the new Conclave API (#366)
AttestationTest required a way to create a separate handler structure to the one required by the public API, so InternalEnclave was introduced to allow a test to setup a custom handler structure.

The only remaining use of the old Enclave class is by the new one itself. This will be cleaned up in a seperate commit.