Kotlin shaded into Conclave SDK (#457)
To avoid versioning issues with Conclave apps using a different version of Kotlin to the one we use, our dependency to it has been shaded into the conclave-mail jar. As the root of the Conclave module hierarchy only this module needs to have the Kotlin packages shaded; the other modules use this shaded version.

Shading Kotlin incorrectly can break Conclave apps written in Kotlin. So to make sure this isn't the case a new class of tests have been introduced, integration tests, which use the published Conclave SDK rather than local project modules.

A consequence of shading though is that any Kotlin language features in the API are lost. So things like typealiases and extension methods no longer work. This isn't a big issue as the public API is designed with Java primarily in mind.