CON-114: CPUSVN added as internal mail header (#470)
This allows enclaves to read mail across microcode updates and to prevent mail encrypted on newer firmware from being decrypted on older firmware.

To avoid reading the header twice, MailPrologueInputStream has been merged into MailDecryptingStream with the private key made optional. This allows the header to be read in once and authenticated (and the stream decrypted) once the private key has been derived.

The EncryptedEnclaveMail typealias has been removed as typealises in the Conclave API are no longer visible in client Kotlin code due to the shading of Kotlin.