enclave-jvm-tester converted to DJVM integration tests (#481)
This is to fix the broken JVM Tester CI pipelines, which broke when Kotlin and Jackson were shaded into the SDK.

The following modules changes were made when moving from enclave-jvm-tester to integration-tests, to simplify things and to make sure they are all under the djvm: module path:

* djvm:djvm-tests, djvm:djvm-tests-auxiliary, djvm:djvm-asserters and djvm:djvm-proto have been merged into djvm:djvm-sandbox-test-bundle
* tests:maths has been renamed to djvm:djvm-simple-test-bundle
* djvm:djvm-test-utils and api have been merged into djvm:djvm-base
* enclave and host have been renamed to djvm:djvm-enclave and djvm:djvm-host respectively

In addition, enclave-utils from the main project has been merged into conclave-enclave as internal API as it wasn't not part of the SDK and so couldn't to be used within integration tests.

The concept of slow integration tests has been introduced, which are tests tagged with “slow”, to prevent the slow sandbox tests running on the PR builds. The JVM Tester pipelines have been changed to be generic Slow Integration test pipelines, which allows other slow integration tests to be added in the future.