Reduce visibility of Kotlin internal in the public API (#486)
The shading of Kotlin exposed any internal members in the API to apps written in Kotlin. This has been fixed by not shading the kotlin.Metadata annotation. For Java apps the members have been renamed to include a space, which is an invalid Java identifier character, to make it harder to access them.