Enclave checks the attestation it receives from host is correct (#507)
Otherwise it was possible for the host to send it an attestation for an entirely different enclave or an older one, etc.

Also, the equals/hashCode for EnclaveInfo has been fixed to include the other properties. Relying on just the code hash is incorrect as changing the revocation level, for example, doesn't change the code hash.