[CON-128] Add Jimfs v1.1 as conclave-filesystem
Unmodified [Jimfs v1.1](https://github.com/google/jimfs/tree/v1.1) cloned to `conclave-filesystem`, which is a gradle module on which `conclave-enclave` and `graal` depend on. I am still opting for a separate module rather than integrating it with an existing one so we can reduce the amount of code graal has to depend on.

Following up is #538, which renames the packages and removes the parent project. The goal is to have only the relevant changes being highlighted in #529.