CON-140: Mail minSize implemented
The format of the stream packet has changed so that padding bytes of zeros are appended to the end of the user bytes. That's then followed by the size of the user bytes and all three sections are encrypted together. To reach the required min size the last packet with user data is padded (if there's space) and then packets with just padding are written, as many as needed. The terminator packet is now one with zero users bytes and zero padding.