CON-140: Introducing PostOffice, which replaces MutableMail
`MutableMail` requires the user to manually increment the sequence number for ordering. However this is easy to forget and not actually consistent with the name - “MutableMail” implies building a single mail object rather than a steam of them. It’s been replaced by `PostOffice` which automatically applies an incrementing sequence number to each call of `encryptMail`.

Inside the enclave it’s important that the enclave’s encryption key be used and not anything else, so the `Enclave.postOffice` methods return instances with the key already applied, and to ensure the enclave key isn’t leaked, these `EnclavePostOffice` instances don’t have a getter on the private key. Further, it’s not possible to create the general `PostOffice` manually inside the enclave to prevent the user from accidentally using a random key.

To make it easier to create a private key, Curve25519KeyPairGenerator has been replaced by `Curve25519PrivateKey.random()` which returns a random private key. The user doesn’t really need the public key and so returning just the private key simplifies things. The public key is already available via a property.

Other changes include moving `decryptMail` from EnclaveInstanceInfo to PostOffice which means one less parameter for the user to pass in, and improving the exception messages to help guide the user when they use the API incorrectly.

The docsite hasn't been updated. That will be done as a follow-up.