Update docker images version, maintainer and remove duplicate code
The `LABEL version` seems to be just an image property and is unrelated with the tag. I think it would be nice to match it with the Conclave version to avoid misunderstandings.

`publish_docker_images.sh` contained duplicate code since part of it has been included in `ci_build_common.sh`. This would cause the publish docker images TC pipeline to build the images twice.