Clear SDK repo directory and build files during incremental builds of SDK
The TC build scripts call for the SDK target to be built multiple times, either directly or through dependencies. Each time it is targeted it publishes new artefacts with new IDs to the maven repository, even if the artefacts have not changed.

In addition, the SDK build folder is used to build the hello-world sample at some point during script execution. The build folders from this are left behind and packed into the distribution zip.

This PR is for a quick-fix rather than properly addressing the issues. Each time the `sdkfiles` task is executed, it clears the maven repo making way for only the new files and it forcibly deletes the hello-world build files.