Add support for file modification time
Implements `__xstat64` for file modification time.
The tests are not exhaustive yet and more will be added once the proposed design is reviewed.

I've introduced a set of tests under `integration-tests/substratevm-jni`, with the objective of unit testing the C/C++ part of the POSIX APIs, unlike the unit tests in `conclave-enclave` which focus on the Java side of the implementation and run on the host's JVM, using mocks to avoid hitting Graal functionality and allowing for rapid iteration.

The tests introduced in this PR aim at testing the C/C++ side, as well as the Graal functionality, for example C type conversion between C and Java, useful for edge case scenarios and catching type mismatching. The nature of these tests mean that the foundation for JNI support is also being set, which will be useful for CON-99.

These tests take advantage of Graal's C interface, some of which is explained in [this tutorial](https://github.com/oracle/graal/blob/release/graal-vm/20.3/substratevm/src/com.oracle.svm.tutorial/src/com/oracle/svm/tutorial/CInterfaceTutorial.java), and include memory allocation (heap and stack), calls to C functions and how to pass C `struct`s between C and Java.

The C `struct`s which need to be exposed for supporting `__xstat64` are now declared in files under `cpp/jvm-enclave-common/include/public`. The `public` directory is part of a new module which is unpacked by our gradle plugin which also passes it to `NativeImage` for include path resolution once `native-image` gets invoked.

To implement these tests I have decided to explore [kotlin polymorphic serialization](https://github.com/Kotlin/kotlinx.serialization/blob/master/docs/json.md#content-based-polymorphic-deserialization) in order to achieve some kind of RPC, without having to manually distinguish between the tests cases.
As such, the `substratevm-jni:common` includes the serializable messages to pass the enclave, containing the test inputs, but also the test code (a call to the POSIX function to be tested), which then returns a serialized return value to be asserted by the host.

A tutorial on JNI support would look very similar to these tests but would focus on user provided `struct`s, our gradle plugin's options (include and library paths) and the compilation process of the C/C++ code (using Conclave provided headers).