CON-248: Migrate native "unit" tests to integration-tests
This migration is to enable enclave tests (conclave-enclave/src/test/kotlin/com/r3/conclave/enclave/EnclaveTest.kt) to run under GraalVM native image.
Also, Java tests, run by AvianTestSuite, are using native image as well. These tests are not migrated to 'messaging' style yet.

As we are now going to build fewer enclave images, it is expected that, with increased number of tests, the build/run time for the tests will be reduces. I am also utilising a polymorphic serialisation to communicate between unit test(s) running on host and the logic running within an enclave.  

The basic structure (sub-projects) of  the test project is:
common <--  this is where `messages` are defined; a message carries the actual logic of the test.
enclave <-- a Testclave - an enclave that handles deserialisation of the messages and serialisation of the responses.
host <-- that is where the unit tests are residing.

Simplest to follow are Echo and Increment messages.