Fix volatile status of shared memory and ensure time only runs forwards.
Occasionally it has been observed in our unit tests that time appears to run backwards, in that the time reported by an enclave is sometimes less than a previous time.

The enclave gets its time from an area of shared memory which is configured when the host first starts the enclave. The enclave treats the memory area as volatile, treating it like it would a memory mapped device. The host writes to this area of memory periodically to update the time and the enclave reads from the same area when it wants to read the time.

Looking at the code, the variable within the structure was defined as volatile but the pointer to the actual memory area should have been treated as volatile. This has been fixed and may actually fix the problems that were observed. However, an additional step has been introduced whereby the enclave now does not return values for time that are earlier than the previous value. The docs have been updated to reflect this.

In addition, it was noticed that the C++ unit tests did not seem to be integrated into the build. A quick implementation has been added here where the unit test CMake target is invoked after building the rest of the C++ code.