[CON-311][CON-316][CON-290][CON-138][CON-79] Improve CI
Implementation proposal for #614 

Currently on TeamCity there are several configurations prefixed with `Rui`, which mostly respect the design proposal from #614, the exceptions being shorcuts taken to test the entire pipeline more quickly and some tasks related with the `master` branch, which can only be properly tested once merged.

Docker images are built by relying on the agents's docker version, rather than the circular dependency of building an image inside a container. This allowed for the removal of the `docker-ce` aptitude package, reducing the size of the image. Two images are built and published as TeamCity artifacts: `aesmd` and `sgxjvm-build`.

No more `docker prune`. Images will not be deleted from the agent's CI caches, which will allow CI to produce docker images much faster. We should manually update the images as explained on #614. We can also introduce a CI option to force an image to be rebuilt for the sake of updating aptitude packages without updating the base image.

`sgxjvm-devenv` has been merged with `sgxjvm-build`.

Most of `devenv.sh` has been merged with `ci_build_common.sh`, meaning most of the docker options are now the same between the two environments.

Environment variables have been introduced to control whether the docker images are built, loaded, saved or pulled, giving us flexibility on our devenvs. By default, devenv will build a new image from the local `Dockerfile`s. This is useful to ensure we are running the appropriate container for the checked out branch. If a container is already running, it will reuse the container, similar to the current behaviour. Rebuilding shoudn't add much overhead since it will use the local docker cache.

Slow integration tests have been fixed and normal integration tests are now building against the SDK zip rather than the locally built SDK triggered by the hardware unit tests. Integration tests have been moved to their own TC configurations.
The reason for the failure was that the SDK zip contains a `conclave-VERSION` directory on the root these days, rather than the actual contents and the gradle settings were picking up the locally built SDK instead.

`PROJECT_VERSION` has been replaced with `USE_MAVEN_REPO`, which can take `artifactory` or `sdk` as a value and only controls which dependencies are the `samples` going to use. When unset, the composite build is used.

Publishing to artifactory is still missing from the configurations, but the scripts have been provided.
Builds of `master` publish as `MAJOR.minor-SNAPSHOT` to artifactory and `latest` for the docker images on the nightly repos.
For builds of `release/` branches I propose the introduction of a CI env var `OBLIVIUM_CONTAINER_RELEASE_REGISTRY_URL` pointing to the docker release repo, set on the "Publish" TC configuration. For the maven artifacts I believe the easiest way would be to override `OBLIVIUM_MAVEN_REPOSITORY` and make it point to the release version.

If we need to rebuild/maintain the `1.0` release, I'd suggest cherry picking this PR and continue on the updated TC configurations. If not possible, we should be able to replicate the configurations using the version controlled TC configurations, something we should track for future releases.

I will review and update the internal-docs if we accept the current proposals.

The PR will remain in draft until accepted, which will then require an update to TC configurations meaning normal TC service will be interrupted while the reconfiguration happens. This PR's gates will fail until the reconfiguration happens and dev branches will need to rebased once the merge happens.