devenv_shell.sh script fix
As noticed, `~/.oblivium_credentials.sh` was out-of-sequence, it should be called before `ci_build_common.sh`.
Also, it was noticed that `GRADLE_OPTS` was empty, and it should in fact use `=-Dorg.gradle.workers.max=$num_cpus --stacktrace`.
That caused another problem: bash was "sending down" to `docker run` `--stacktrace` as an individual parameter. 
There was no easy fix but to use an array of strings (for the parameters) or repeat the `-Dorg.gradle...` argument once again.
To avoid code repetition, and also because it seems to be more "elegant", the array of strings was chosen for this solution.
I was careful to check the entire codebase to see whether those changes would affect other scripts. And apparently, they don't.

One final change was to disable the `x` flag in the devenv_shell.sh script to provide a better output when the script was in its final stage.