CON-344: Introduce a new 'mock' configuration for building enclaves.
The plugin has been modified to support a new configuration named 'mock'. This is very similar to the functionality provided by the mock host but instead of having to change your host code and host build configuration to load the enclave as a mock, you just have to depend on the enclave 'mock' artifact.

So for the hello-world sample, to use the mock all you need to do is:

`./gradlew build -PenclaveMode=mock`

In order to support this, a number of the mock classes needed to be moved from `testing/internal` to `enclave/internal` and `host/internal`. This does not affect the standard mock host mode in any way: this can still be used with no API changes.

In the host itself the enclave handle is managed by the `LocalEnclaveHandle` class. The reason this is not called `MockEnclaveHandle` is to prevent confusion/conflicts with the `MockEnclaveHandle` in the testing package.

I've updated all the documentation to cover the new enclave mode. Hopefully I've caught all the necessary changes.