Prompt user to refer to docs if external signed material is not present.
Implementation of CON-326. Fixes a usability issue when using external keys to sign enclaves. The process is two-step when using an external key so if a customer attempts to build their enclave without first preparing the signing material and signing the enclave then the build fails with a non-obvious message.

A better error message has been provided including a link to the relevant part of the documentation on how to sign enclaves.