CON-396 CON-397 Remove MockHost and replace with the mock enclave mode loaded with EnclaveHost.load()
This PR originally was named "Add 'enclave' property to EnclaveHost in Mock mode to allow unit tests to access enclave state." and was the prep work for removing MockHost. However, during the PR review it didn't make sense to merge the PR changes without actually removing the MockHost class. So, this PR includes the original change plus all the changes required to remove the original MockHost.

A summary of the changes:
* Remove the MockHost class (obviously!)
* Remove the conclave-testing package as it only contained a couple of support classes and tests that could live elsewhere
* Add the 'enclave' property to EnclaveHost. If code attempts to access the enclave property in any mode other than Mock than `IllegalStateException` is thrown.
* Update all documentation and release notes
* Update the sample to use the new mock unit test method

Note that one important side-effect of the approach that is used to load mock enclaves is that if an enclave is on the classpath then it can be loaded directly via `EnclaveHost.load()`. This is really useful for unit tests within an enclave project because you just need to add a test dependency on `conclave-host` then use `EnclaveHost.load()` to get a mock enclave. However, you need to be careful when working with non-mock enclaves because if the enclave class is on the classpath as well as the enclave file present in the resources then you will get a `multiple enclaves found` exception.