Upgrade Dokka to latest version and fix a few documentation issues
This is a major overhaul of our dokka documentation generation and involves the following:
* Update to the latest dokka version
* Fix a number of blocking issues in the current dokka version by forking it
* Include our own custom build of dokka in the build container
* Generate HTML documentation instead of javadoc (new feature of dokka)

The upgrade was not straightforward. There were a number of bugs in the new version (some present in the old version). The fixes to these in dokka were not that difficult but would be hard to get merged upstream due to the amount of tests that would need to be written and possible arguments over the correct solution. So instead, we now have an R3 Conclave dokka fork of the repo that includes a branch with the patches. There is a build script in the /dokka folder that will generate a new version and zip it. This then needs to be uploaded to our fork as a release in github. Instructions in the README.md in the same folder.

Our build container pulls the release from github and puts a local maven repository in /opt/dokka. This is where the dokka plugin is loaded from and is required to be present otherwise our root `build.gradle` fails (hence we can't build dokka as part of our normal build process).