CON-385: Fix for potential data race condition in EnclaveShareData's real_time()
There's a potential racing condition in the following (other parts need to be verified too):

```
uint64_t EnclaveSharedData::real_time() {​​​​​​​​
 init();
 SharedData sd;
 getSharedData(sd);
 // Ensure time cannot go backwards
 if (sd.real_time > last_time_) {​​​​​​​​
 last_time_ = sd.real_time;
 }​​​​​​​​
 return last_time_;
}​​​​​​​​
```

Due to standard locks not scaling properly, I've implemented a Double-checked lock pattern.
(To understand the logic behind it, please refer to https://preshing.com/20130930/double-checked-locking-is-fixed-in-cpp11)

I've also did a small change on the "Enclave Shared Data" native tests. `EXPECT_EQ` provides the information within the parameters, which is useful in case of failure.

Also did some small improvements, such as using const reference instead of direct use of pointer (this saves some time as the function doesn't need to verify if the pointer is a `nullptr`). "const" to methods that don't affect / change its class members.