CON-384 Tidy up enclave configuration console output
Before this commit, the output generated by any invocation of
project.exec from Gradle would be written to the console without
taking into account the log flags used. This means the user could not
suppress any output generated by the project.exec call.

The fix was to redirect the output generated by the project.exec call
and store it in a stream. Once the project.exec call is
completed, the output stored in the stream is printed to the console using the
project.logger. By using the project.logger any log flag (--quiet,
--warn, --info, --debug) is taken into consideration.