CON-318 Include the version and git hash in MANIFEST.MF
Updated the build.gradle file in the root directory of the sgxjvm project. The changes add the Conclave-Release-Version, and Conclave-Revision attributes to the manifest files of all jar files generated by the project using the gradle task jar. This covers at least all the jars created for the sdk.

The attribute Conclave-Release-Version corresponds to the conclave version.
The attribute Conclave-Revision corresponds to the git hash of the last commit when the jar was created.

Info log messages were added to keep track of conclave release version, and the conclave revision used while building the jars. To see the logs use the --info flag from gradle.

A warning log message was added which is written when gradle fails to get the git hash of the last commit.