CON-346: Fix IDE docs integration for VS Code and Eclipse.
This fixes a couple of issues. Unfortunately none of them make Conclave javadocs work with IntelliJ out of the box. However, with these changes, javadocs work with both VS Code and Eclipse, albeit with a few formatting issues which are a result of bugs in Dokka and will need to be addressed outside the scope of this PR.

Firstly, our modified dokka has been updated to include a fix that completely broke the ability for IDEs to detect classes and methods. Our build container has been updated to pull this new version.

Secondly, the hello world sample has been updated to include docs configuration for eclipse as well as the existing Idea configuration. VS Code shares the same eclipse configuration so is also enabled by this.

Lastly, the previous change to switch from Javadoc to HTML documentation meant the maven javadoc artifact contained the wrong format data. Our build now builds both HTML and javadoc and packages them correctly.