FEAT-754: Adding cert-based identity to CorDapp sample
Part of FEAT-754, adds:
- Private/secure mailer identity validation
- Initial Enclave Flow API

Conclave Identity goals: 
- to assert the other party’s identity securely, with no trust in the enclave’s host
- to assert the other party’s identity privately, without letting the host know whose identity we're verifying

To fulfil the above identity goals, we operate as follows:
- the sender encodes its certificate path, which includes the sender as the Subject, the Corda trusted root certificate and any intermediate certificate in the chain of trust
- the sender signs a secret that is only known to the sender and the enclave
- the sender creates a mail with a body containing the encoded certificate path and the signature bytes from the 2 points above, plus the sender’s public key
- the mail is then encrypted and sent to the enclave
- the enclave decrypts the mail and verifies the signature bytes against the authenticated sender key using the sender public key
- the enclave verifies the sender's certificate path validity using the certificate authority’s public key that is embedded in the enclave itself (trustedroot.cer), and that the public key used to verify the signed secret belongs to the certificate path Subject
- the enclave knows the sender’s name and public key that define the sender’s identity and that the identity is authentic

Rather than sending the identity with every mail, it makes sense to define a session within which an identity is valid. For this the authenticated sender key is used, as it is valid while the same encryption key is used to encrypte/decrypt messages between the sender and the enclave. A side effect of this is that identities can keep accumulating in the cache, therefore in a future iteration this will need to be taken into account.

The identity is therefore sent by the initiator flow with the first message that is sent to the receiver and is implemented by a dedicated mail with a special "--concave-login" topic and a payload containing the serialized MailerIdentity instance prepared by the sender. The Enclave handles mails with a topic starting with "--conclave-" as a special case message.

Once an identity is verified and added to the Enclave identities cache, the EnclaveMail that is handed to the receiveMail method (for the whole session) will contain a verified MailerIdentity instance that can be retrieved with the getAuthenticatedSenderIdentity() method and offers 3 properties:
- isAnonymous() - a boolean property indicating whether the sender sent the identity or not
- getName() - the X509 name of the sender or "anonymous" if the sender did not send its identity
- getPublicKey() - the X509 public key of the sender or the encryption public key if the sender did not send its identity

If the identity was not sent the getAuthenticatedSenderIdentity() will return an instance of the AnonymousSender. In the case the enclave fails to validation the sender identity, the sending flow will get a flow exception with an error message about the authentication failure.

replaces PRs https://github.com/corda/sgxjvm/pull/618 and https://github.com/corda/sgxjvm/pull/624

TO DOs
- update/add comments to some classes and methods
- update cordapp docs to reflect refactoring changes (waiting for code review)