Topic doc / code comment #changes.
As requested in CON-369, Changes were made to explain better the use of topic, and for the information to be consistent amongst the documentation.

"In https://docs.conclave.net/writing-hello-world.html#mail-headers we say "Topics are scoped per-sender and are not global. The client can send multiple streams of related mail by using a different topic for each stream, and it can do this concurrently." but in the javadocs for EnclaveMailHeader.getTopic we say "The topic is not parsed by Conclave and should be unique across time to avoid replay attacks:". These should be updated to be consistent.

In the javadocs for EnclaveMailHeader.getTopic we also mention that the topic could be used for load balancing, but this will not be possible until we have enclave clustering, so we should remove this statement for the time being."