CON-459: Separate topic no longer used for login in CorDapp sample
**Issue summary**
The use of a separate topic for the login is problematic. The host can trivially choose not to pass on these login mail messages, and if the enclave is written such that the remote client cannot detect if it's being treated as a logged in user or anonymous user then it is no way of knowing if the host decides to not pass on the login message. Rather what needs to happen is that the login message needs to be the first message of the user's topic, thus protecting against the host dropping the mail.

**Solution implemented**
The code was updated to guaranty that the login message is always the first message per topic to be sent to the enclave. This guaranties that the enclave throws an error next time it receives a message if the host decides to drop the login message.

**Implementation details**
 1) The first byte of the login message indicates whether the user wants to authenticate or not. One indicates the user wants to remain anonymous while zero indicates the user wants to be authenticated. If the first byte is one then the following bytes will be set to zero. The message is padded with zeros to prevent people from gaining insights based on the login message size. The number of bytes padded to the message varies in size and takes into account the average size of an identity object. Again this is to prevent statistical analyses insights. If the first byte is zero then the following bytes represent the identity of the user.

**Other changes**
 1) ReverseFlow now receives a fourth parameter. The anonymous parameter allows the user to run the flow as an anonymous user if set to true.

 2) A new test case was added to check if the flow runs as expected if the user decides to remain anonymous

**Remarks**
1) Keep in mind that the first enclave mail of a given topic will always be zero.