CON-522: Use LinuxExec to execute sign_tool in a docker container on Windows
During the release of Conclave V1.1 a licensing issue was found with the Windows build of sgx_sign.exe. This PR removes the tool from the SDK and instead uses the existing LinuxExec class to run the native Linux version inside a docker container (like we do for native-image).

Ideally, the macOS version of the tool should also be removed for consistency, using LinuxExec on macOS too. However, this PR will be merged into Conclave V1.1 at quite a late stage of the release cycle therefore to reduce risk of introducing a bug, this PR only affects Windows builds and goes so far as to include 'is_windows()' conditions, invoking the original code if not on Windows.

When merged back into the main branch, the conditions should probably be removed, always using LinuxExec on every platform.

In addition, it  was noticed that ld.exe on Windows also has the same licensing issue. This PR removes that executable from the SDK which means that Avian builds on Windows will fail. Conclave 1.1 deprecates Avian so this was deemed acceptable. A note has been added to the deprecation notice in the release notes.