Update sgxjvm-build and aesmd images versions to 1.2
This impacts the publishing on artifactory. Since we left it as 1.0 means that the latest 1.0 on artifactory is now of a 1.1 image. The actual 1.0 is probably still there, since artifactory seems to accumulate some images. If we really needed it we can do some archaeology.

I have manually uploaded the correct images to artifactory with the 1.1 tag.