Re-add a modified  Outliers.kt, to compile integration tests
As Outliers.kt relied on Avian code, I have removed it.
But this seems to create issue in the compilation.
Here we fix it, but we need further attention to integration tests
as they are currently failing. We might need to ignore some of them
as not currently relevant to GraalVM