Fix SealingTest by using EnclaveHost mechanism
Previously SealingTest.kt was a test using non-public Enclave
functions.
This was relying on Avian, which was faster than GraalVM and easier to
embed in a test.
As we moved to Graal, the way of bundling Graal is different from
Avian and therefore we switch to another way of testing, specifically
we create the Enclave as if we were a client, i.e. using EnclaveHost
and related functions. This is similar to what we do in QA test apps.