Make MockEnclaveEnvironmentHardwareCompatibilityTest.kt to run in Debug mode only.
Fix configs so that the test is ignored when running integration tests in Simulation mode
and make this test run only in debug mode.
Also, when generating the combinations for the enclaves to test, we specify only one combination coming
out from IsvSvn and IsvProdId. This is because in GraalVM we do not have a mechanism to create
dynamic enclaves on the fly as we had in Avian.
We postpone to a later stage the decision on how to make this test more complete as it was before,
i.e. how to test all the combinations as it was before.
The test has been put back in the tests module, there isn't now any need to have a separate module for it
as we run it only in Debug mode