CON-462 - Upgrade GraalVM to 21.2
Upgraded GraalVM from 21.1 to 21.2

Upgraded MX version from 5.294.1 to 5.301.0 because the new version of GraalVM requires mx version to be at least version 5.301.0.

MX is now fetched using git instead of curl. An error is thrown while building Graal. The error is thrown because MX_HOME directory must contain the dot folders related to git. The dot folders are created while using the git command. The archive fetched using curl does not have all the dot folders related to git like .git. This error only appeared after upgrading Graal and mx build tool. The error contains the message "TypeError: argument of type 'NoneType' is not iterable". This is because VC.get_vc_root( ... ) returns None when the argument is /opt/$MX_HOME. VC stands for version control and the function get_vc_root returns the path to the directory that contains the dot folders related to git inside the path or subpath of /opt/$MX_HOME/. Otherwise, it returns None.

Added the patch file graal_21.2.patch.
The patch graal_21.2.patch is almost identical to graal_21.1.patch. The patch graal_21.2.patch was updated due to a conflict with file suite.py when the patch was applied to GraalVM 21.2.

Added stubs for __strtok_r, and nanosleep to fix the "Hello World" sample build.

Cleaned the ./graal/patch directory. It is not necessary to have multiple patches. Firstly, Conclave always uses the latest version of the patch. Secondly, git already keeps track of the evolution of the patch. Summary of changes:
* Updated clone.sh to point to the new patch file name.
* Renamed graal_21.2.patch to graal_patch.
* Deleted the older patches.
* Updated the README file to reflect the changes.

Updated how to upgrade Graal in the internal-docs.

Updated the docs that talk about javascript-python. The example provided for Python had a compilation error. Renamed a function to minimize confusion while copying and pasting the code.

Considering the effort it requires to upgrade GraalVM, it  incredibly unlikely a developer will be able to simply use a different version of Graal. Thus, the property variable graalVMSDKVersion was removed from GradleEnclavePlugin to prevent developers from changing GraalVM version.