CON-482 Provide startup and shutdown hooks for the Enclave
Added onStartup and onShutdown hooks. The onStartup hook is called after the enclave is created and initialised but before the enclave runs. The onShutdown is called just before the enclave is destroyed. The first hook allows developers to add extra configuration to the enclave before it starts running while the second enclave gives an opportunity to the enclave to release all resources used by the enclave in a deterministic manner.

The hooks are available in mock and non-mock mode.

Added unit tests to exercise the hooks while using a mocked host.

# Changes

# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?