CON-95 API to update the remote attestation
Added a method to EnclaveHost which performs a fresh attestation and updates the enclaveInstanceInfo property. The attestation is still performed on startup but it can now be performed as many times as required by calling the method Enclavehost.runAttestation.

As part of this work, parts of the code related to the attestation were refactored. The logic was not change but some blocks of code were moved to their own classes and all attestation related files grouped inside their package.


# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?